+++ src/u8c/end.c
@@ -19,7 +19,7 @@
 # include <stdint.h>
 # include <stdlib.h>
 # include <u8c/end.h>
-# include <u8c/freeu32.h>
+# include <u8c/u32free.h>
 # if defined(u8c_bethrdsafe)
 # include <threads.h>
 # endif
@@ -30,7 +30,7 @@ uint_least8_t u8c_end(void) {
 # if defined(u8c_bethrdsafe)
 	mtx_destroy(&u8c_errlock);
 # endif
-	u8c_freeu32(u8c_err);
+	u8c_u32free(u8c_err);
 	u8c_stat = UINT8_C(0x1);
 	return UINT8_C(0x0);
 }