+++ src/u8c/vprint.c
@@ -18,10 +18,10 @@
 # include <stdint.h>
 # include <stdio.h>
 # include <stdlib.h>
-# include <u8c/freeu32.h>
-# include <u8c/freeu8.h>
 # include <u8c/seterr.h>
+# include <u8c/u32free.h>
 # include <u8c/u8enc.h>
+# include <u8c/u8free.h>
 # include <u8c/SIZE_C.h>
 # include <u8c/vfmt.h>
 # include <u8c/vprint.h>
@@ -37,7 +37,7 @@ uint_least8_t u8c_vprint(FILE * _fp,uint_least32_t const * const _msg,va_list _a
 		u8c_seterr((uint_least32_t[]){UINT32_C(0x75),UINT32_C(0x38),UINT32_C(0x63),UINT32_C(0x5F),UINT32_C(0x76),UINT32_C(0x70),UINT32_C(0x72),UINT32_C(0x69),UINT32_C(0x6E),UINT32_C(0x74),UINT32_C(0x3A),UINT32_C(0x20),UINT32_C(0x66),UINT32_C(0x77),UINT32_C(0x72),UINT32_C(0x69),UINT32_C(0x74),UINT32_C(0x65),UINT32_C(0x3A),UINT32_C(0x20),UINT32_C(0x55),UINT32_C(0x6E),UINT32_C(0x61),UINT32_C(0x62),UINT32_C(0x6C),UINT32_C(0x65),UINT32_C(0x20),UINT32_C(0x74),UINT32_C(0x6F),UINT32_C(0x20),UINT32_C(0x77),UINT32_C(0x72),UINT32_C(0x69),UINT32_C(0x74),UINT32_C(0x65),UINT32_C(0x20),UINT32_C(0x74),UINT32_C(0x6F),UINT32_C(0x20),UINT32_C(0x73),UINT32_C(0x74),UINT32_C(0x64),UINT32_C(0x6F),UINT32_C(0x75),UINT32_C(0x74),UINT32_C(0x2E),UINT32_C(0x0),}); /* u8c_vprint: fwrite: Unable to write to stdout. */
 		return UINT8_C(0x1);
 	}
-	u8c_freeu32(str0);
-	u8c_freeu8(str1);
+	u8c_u32free(str0);
+	u8c_u8free(str1);
 	return UINT8_C(0x0);
 }