+++ changelog.md
@@ -1,3 +1,9 @@
+# 14
+
+* Free error message when `u8c_geterr` is called (after copying, of course).
+* Update `u8c_freeu8`.
+* Rename `u8c_freeu32` to and `u8c_u32free` and `u8c_freeu8` to `u8c_u8free`.
+
 # 13
 
 * Fix `u8c_txt` in C++.
@@ -12,7 +18,7 @@
 * Update Makefile.
 * Use constant variables more.
 * Create macro for creating human-readable UTF-32 strings; `u8c_txt`.
-* Add macros for deallocating UTF-32 and UTF-8 strings (use these instead of `free` og `std::free`); `u8c_freeu32` and `u8c_freeu8`.
+* Add macros for deallocating UTF-32 and UTF-8 strings (use these instead of `free` or `std::free`); `u8c_u32free` and `u8c_u8free`.
 * Optimisations.
 
 # 10